run1/public_service/requirements.txt
fastapi
uvicorn
httpx
google-auth
requests

